It passes tests (clang edited)

+++ tasks/shemetov_d_increasing_contrast/mpi/src/ops_mpi.cpp
@@ -3,6 +3,7 @@
 #include <mpi.h>
 
 #include <algorithm>
+#include <cstddef>
 #include <cstdint>
 #include <vector>
 
@@ -52,7 +53,7 @@ bool IncreaseContrastTaskMPI::RunImpl() {
   constexpr float kFactor = 1.3F;
 
   for (size_t i = 0; i < local_in.size(); ++i) {
-    const int v = static_cast<int>(local_in[i] * kFactor);
+    const int v = static_cast<int>(static_cast<float>(local_in[i]) * kFactor);
     local_out[i] = static_cast<uint8_t>(std::clamp(v, 0, 255));
   }
