+++ tasks/shemetov_d_increasing_contrast/seq/src/ops_seq.cpp
@@ -1,6 +1,7 @@
 #include "shemetov_d_increasing_contrast/seq/include/ops_seq.hpp"
 
 #include <algorithm>
+#include <cstddef>
 #include <cstdint>
 
 #include "shemetov_d_increasing_contrast/common/include/common.hpp"
@@ -25,7 +26,7 @@ bool IncreaseContrastTaskSEQ::RunImpl() {
   constexpr float kFactor = 1.3F;
 
   for (size_t i = 0; i < GetInput().size(); ++i) {
-    const int value = static_cast<int>(GetInput()[i] * kFactor);
+    const int value = static_cast<int>(static_cast<float>(GetInput()[i]) * kFactor);
     GetOutput()[i] = static_cast<uint8_t>(std::clamp(value, 0, 255));
   }
